AuraPlay + wallet connection: readiness audit, runtime wiring, test/doc updates
Add the AuraPlay Production-Readiness Audit and refresh the AuraPlay gap
docs. Wire wallet-connection feature/runtime changes across the app shell
(MainAura/MainTab, WalletConnection, AuraPlay playback runtime/orchestration)
and MusicFeature discovery/persistence. Drop the obsolete
AuraPlaySchemaVersions and AuraPlayMigration test; add CapturingAuraPlayLogger
and expand audio-engine, NFTKit, ProviderKit, and Auralis test coverage.

Excludes the AutopilotCore/PlannerCore POC packages and the embedded
WalletConnectorDemo app repo.

+++ Auralis/NFTKit/Tests/NFTKitTests/NFTMetadataUpdaterTests.swift
@@ -237,7 +237,7 @@ struct NFTMetadataUpdaterTests {
         NFTMetadataUpdater.updateNFTFromMetadata(
             nft: nft,
             metadata: [
-                "seller_fee_basis_points": .number(750)
+                "seller_fee_basis_points": .int(750)
             ]
         )
